requirements.txt
ucimlrepo
pandas
scikit-learn
matplotlib
seaborn
numpy
scipy
statsmodels
torch
torchvision
torchaudio
distinctipy
jupyterlab
pyyaml
tqdm
joblib
